[BUGFIX beta] Ensure @controller passes args correctly
@controller was not passing args correctly for parameter-less decorator
usage. We weren't actually testing it, which is why we didn't catch the
issue. This adds tests for both @service and @controller and fixes the problem.

(cherry picked from commit 53e6a9cec0990bbd586b4c06f9932d7d7aaf92ff)
